+++ lib/djvCore/Math.cpp
@@ -35,8 +35,8 @@ namespace djv
             int getRandom(int value)
             {
                 std::uniform_int_distribution<uint32_t> uint_dist;
-                const float r = uint_dist(rng) / static_cast<float>(uint_dist.max());
-                return static_cast<int>((value + 1) * r);
+                const float r = static_cast<float>(uint_dist(rng)) / static_cast<float>(uint_dist.max());
+                return static_cast<int>(static_cast<float>(value + 1) * r);
             }
 
             float getRandom(float min, float max)
@@ -47,8 +47,8 @@ namespace djv
             int getRandom(int min, int max)
             {
                 std::uniform_int_distribution<uint32_t> uint_dist;
-                const float r = uint_dist(rng) / static_cast<float>(uint_dist.max());
-                return min + static_cast<int>((max - min + 1) * r);
+                const float r = static_cast<float>(uint_dist(rng)) / static_cast<float>(uint_dist.max());
+                return min + static_cast<int>(static_cast<float>(max - min + 1) * r);
             }
 
             void setRandomSeed(unsigned int value)
